Updated the fit strategy implemented by proset.utility.select_hyperparameters() to reduce overfitting if using multiple batches.
Squashed commit of the following:

commit e2aa9be0e5b5ccfc414fc86f371eb535f4800951
Author: Nikolaus Ruf <nikolaus.ruf@t-online.de>
Date:   Sat Oct 8 13:09:09 2022 +0800

    Updated README.md.

commit 8ffccfad296e72a12b55684b8fd9ccb1fc45a41b
Author: Nikolaus Ruf <nikolaus.ruf@t-online.de>
Date:   Sat Oct 8 13:05:12 2022 +0800

    Updated PDF included with repo.

commit 88a2d9308d67701ee11f2221062221a7280fac6e
Author: Nikolaus Ruf <nikolaus.ruf@t-online.de>
Date:   Sat Oct 8 13:04:32 2022 +0800

    Updated cross-reference to refer to new algorithm.

commit 013241b9c9ca69a4cd577df39a3a78d00953fd0a
Author: Nikolaus Ruf <nikolaus.ruf@t-online.de>
Date:   Sat Oct 8 12:53:41 2022 +0800

    Updated documentation.

commit 4dce17b84b863ca20feea4b97cb76d0d1d87ae1c
Author: Nikolaus Ruf <nikolaus.ruf@t-online.de>
Date:   Sat Oct 8 10:12:16 2022 +0800

    Updated logging for the objective base class.

commit e524c92ba7fe0e96bf4d7e17917d28e0a8eee9b6
Author: Nikolaus Ruf <nikolaus.ruf@t-online.de>
Date:   Sat Oct 8 10:07:33 2022 +0800

    Updated logging for parameter search.

commit 3a6a47eff8d1e884e858b4d56039a1e6d5bece05
Author: Nikolaus Ruf <nikolaus.ruf@t-online.de>
Date:   Sat Oct 8 10:03:54 2022 +0800

    Updated logging for model class.

commit 354fdc8c5b3d8fbf6f79d0a840152c6c0cd1b1cf
Author: Nikolaus Ruf <nikolaus.ruf@t-online.de>
Date:   Fri Oct 7 18:25:10 2022 +0800

    Updated lambda_v grid for CIFAR-10 benchmark.

commit d274e1f69a101af2c0e421bc2b87e0dd95fcc053
Author: Nikolaus Ruf <nikolaus.ruf@t-online.de>
Date:   Thu Oct 6 20:06:46 2022 +0800

    Parameter search plot now has a second version with lambda_v on the x-axis and lines distinguished by number of batches.

commit 80e28be25b0298dc69eaf766072eaa67c737178c
Author: Nikolaus Ruf <nikolaus.ruf@t-online.de>
Date:   Thu Oct 6 17:28:28 2022 +0800

    Finished reviewing the benchmark scripts.

commit a2b5e918397c7c37cc28579a5ebd5ea4fff85417
Author: Nikolaus Ruf <nikolaus.ruf@t-online.de>
Date:   Thu Oct 6 12:14:17 2022 +0800

    Bugfix: plot function checked for familiarity to be strictly positive but zero is admissible for numerical reasons.

commit b7fd30a696b55e25cd67559f8048719d5b54ea53
Author: Nikolaus Ruf <nikolaus.ruf@t-online.de>
Date:   Sat Sep 24 21:54:37 2022 +0800

    Simplified structure of scripts for first benchmark study as they now deal with only one model each.

commit 4ad69f54ff0058faa816520b81cf6076d16e90b9
Author: Nikolaus Ruf <nikolaus.ruf@t-online.de>
Date:   Sat Sep 24 20:16:19 2022 +0800

    Bugfix: random state for final model was arbitrary, leading to non-reproducable benchmark results.

commit 66b6b9aa6b308cb29a8ec15216341aeeec48813e
Author: Nikolaus Ruf <nikolaus.ruf@t-online.de>
Date:   Sat Sep 24 16:37:35 2022 +0800

    Updated checker rot benchmark case to latest interface.

commit a553f0895b7d22c076d039457d9197d761f7a8b5
Author: Nikolaus Ruf <nikolaus.ruf@t-online.de>
Date:   Sat Sep 24 16:18:51 2022 +0800

    Added second experiment for XOR 6f benchmark as first selects the maximum number of batches as optimal.

commit 7d6227e0f2a2d2b64f36b23b26e296cf6c0b39b3
Author: Nikolaus Ruf <nikolaus.ruf@t-online.de>
Date:   Sat Sep 24 11:51:30 2022 +0800

    Updated XOR 6f benchmark to latest interface.

commit 16789bdd5950d1c1612146df5556f1b2c81bc78c
Author: Nikolaus Ruf <nikolaus.ruf@t-online.de>
Date:   Sat Sep 24 10:54:07 2022 +0800

    Updated checker benchmark to latest interface.

commit 8ea27884598221df0dae907ffb40495ed28c2bbf
Author: Nikolaus Ruf <nikolaus.ruf@t-online.de>
Date:   Tue Sep 20 21:33:10 2022 +0800

    Changed data type specifications for numpy to avoid deprecation warnings during unit testing.

commit 4d3c2de665efb9cbc5755119beb2b777d774bae6
Author: Nikolaus Ruf <nikolaus.ruf@t-online.de>
Date:   Tue Sep 20 21:08:00 2022 +0800

    Updated smaller benchmark cases to use new fit function.

commit e2c3df8e219661ce2006d654f17fe438c8de2fc6
Author: Nikolaus Ruf <nikolaus.ruf@t-online.de>
Date:   Tue Sep 20 08:38:29 2022 +0800

    Updated fit strategy to choose only lambda_v and number of batches in a single stage (code copied from obsolete dev branch).

commit cfe2be2a797220fd27c3e16b8bc57717ad3dba57
Author: Nikolaus Ruf <nikolaus.ruf@t-online.de>
Date:   Tue Sep 20 08:26:37 2022 +0800

    Bugfix: made check for batches with no prototypes more robust (numerical errors could lead to negative weights in rare cases which bypassed the old check).

commit 3344c37babd003f766a8a9b0afb94a4524cfdc60
Author: Nikolaus Ruf <nikolaus.ruf@t-online.de>
Date:   Tue Sep 20 08:24:53 2022 +0800

    Updated package version.

+++ proset/objectives/tf_objective.py
@@ -173,7 +173,7 @@ class TfObjective(Objective, metaclass=ABCMeta):
         """
         return {
             "prototypes": self._sample_data["cand_features"].numpy().astype(**FLOAT_TYPE),
-            "target": self._sample_data["cand_target"].numpy().astype(np.integer),
+            "target": self._sample_data["cand_target"].numpy().astype(int),
             "feature_weights": parameter[:self._meta["num_features"]].astype(**FLOAT_TYPE),
             "prototype_weights": parameter[self._meta["num_features"]:].astype(**FLOAT_TYPE),
             "sample_index": self._sample_data["cand_index"]
